Configure rollup for deployment
