+++ .eslintrc
@@ -1,30 +1,33 @@
 {
-    "extends": "airbnb",
-    "parser": "babel-eslint",
-    "rules": {
-        "id-length": [2, { "exceptions": ["x", "y", "i", "j", "r", "g", "b", "a"] }],
-        "no-unused-vars": [2, { "vars": "all", "args": "none" }],
-        "max-len": [0],
-        "no-loop-func": [0],
-        "no-plusplus": [0],
-        "no-continue": [0],
-        "no-prototype-builtins": [0],
-        "no-underscore-dangle": [0],
-        "jsx-a11y/href-no-hash": [0],
-        "no-restricted-syntax": [0],
-        "arrow-body-style": [0],
-        "object-curly-spacing": [0],
-        "padded-blocks": [0],
-        "no-mixed-operators": [0],
-        "arrow-parens": [0],
-        "prefer-spread": [0],
-        "func-names": [0],
-        "class-methods-use-this": [0],
-        "no-restricted-properties": [0],
-        "import/no-extraneous-dependencies": [0],
-        "import/first": [0],
-        "space-before-function-paren": [0],
-        "consistent-return": [0],
-        "no-for-of-loops/no-for-of-loops": 2
-    }
+  "extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint"],
+  "rules": {
+    "id-length": [2, { "exceptions": ["x", "y", "i", "j", "r", "g", "b", "a"] }],
+    "no-unused-vars": [2, { "vars": "all", "args": "none" }],
+    "max-len": [0],
+    "no-loop-func": [0],
+    "no-plusplus": [0],
+    "no-continue": [0],
+    "no-prototype-builtins": [0],
+    "no-underscore-dangle": [0],
+    "jsx-a11y/href-no-hash": [0],
+    "no-restricted-syntax": [0],
+    "arrow-body-style": [0],
+    "object-curly-spacing": [0],
+    "padded-blocks": [0],
+    "no-mixed-operators": [0],
+    "arrow-parens": [0],
+    "prefer-spread": [0],
+    "func-names": [0],
+    "class-methods-use-this": [0],
+    "no-restricted-properties": [0],
+    "import/no-extraneous-dependencies": [0],
+    "import/first": [0],
+    "space-before-function-paren": [0],
+    "consistent-return": [0],
+    "@typescript-eslint/ban-ts-comment": [0],
+    "@typescript-eslint/no-non-null-assertion": [0],
+    "@typescript-eslint/explicit-module-boundary-types": [0]
+  }
 }
