+++ package.json
@@ -4,15 +4,20 @@
   "author": "David Chanin <chanindav@gmail.com> (http://chanind.github.io/hanzi-writer/)",
   "description": "Hanzi Writer is a free and open-source javascript library for both animating simplified Chinese characters and quizzing users on character stroke order.",
   "repository": "chanind/hanzi-writer",
-  "main": "dist/hanzi-writer.js",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.esm.js",
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-typescript": "^7.10.4",
+    "@rollup/plugin-babel": "^5.2.1",
+    "@rollup/plugin-node-resolve": "^9.0.0",
     "@sinonjs/fake-timers": "^6.0.1",
-    "@types/jest": "^26.0.13",
+    "@types/jest": "^26.0.14",
     "@types/nise": "^1.4.0",
     "@types/sinonjs__fake-timers": "^6.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.1.1",
+    "@typescript-eslint/parser": "^4.1.1",
     "babel-core": "^6.26.3",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.3.0",
@@ -20,22 +25,18 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
     "codecov": "^3.7.2",
-    "eslint": "^7.8.1",
+    "eslint": "^7.9.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
-    "grunt": "^1.3.0",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-uglify": "^5.0.0",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-webpack": "^4.0.2",
     "hanzi-writer-data": "^2.0.1",
     "jest-canvas-mock": "^2.2.0",
     "jest-cli": "^26.4.2",
     "jest-fetch-mock": "^3.0.3",
-    "nise": "^4.0.4",
+    "rollup": "^2.27.0",
+    "rollup-plugin-license": "^2.2.0",
+    "rollup-plugin-terser": "^7.0.2",
     "semantic-release": "^17.1.1",
     "typescript": "^4.0.2",
     "webpack": "^4.44.1"
@@ -43,11 +44,14 @@
   "scripts": {
     "test": "jest",
     "lint-test": "eslint -c .eslintrc-jest src",
-    "build": "grunt",
-    "prepublishOnly": "grunt",
+    "build": "rm -rf dist && rollup -c && tsc --emitDeclarationOnly",
+    "prepublishOnly": "rollup -c",
     "semantic-release": "semantic-release",
     "typecheck": "tsc --noEmit --skipLibCheck"
   },
+  "files": [
+    "dist"
+  ],
   "jest": {
     "setupFiles": [
       "jest-canvas-mock"
