+++ tsconfig.json
@@ -13,6 +13,9 @@
       "DOM",
       "DOM.Iterable"
     ] /* Specify library files to be included in the compilation. */,
+
+    "declaration": true,
+    "declarationDir": "dist/types",
     // "allowJs": true,                       /* Allow javascript files to be compiled. */
     // "checkJs": true,                       /* Report errors in .js files. */
     // "jsx": "preserve",                     /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
@@ -73,5 +76,5 @@
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
   "include": ["src"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "dist"]
 }
